+++ src/main/java/org/ru/skypro/lessons/spring/EmployeeApplication/repository/PositionRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 public interface PositionRepository extends CrudRepository<Position, Integer> {
     Position findByPositionName(String positionName);
-    Position findByPositionId(Integer positionId);
+
+    Position findByPositionId(Long positionId);
 }
